requirements.txt
python-pptx>=1.0.2
pymupdf>=1.25.4
tqdm